refactor(daemon): split storage_decision_report_cmd by concern (831→364)
Pure relocation within package main — no call-site or behavior changes:

- storage_decision_report_cmd.go (364): command entry + report build,
  summarize, warnings, truth derivation
- storage_decision_report_types.go: report DTO/value types
- storage_decision_report_sqlite.go: read-only SQLite query layer
  (column probing, capability + decision-row queries)
- storage_decision_report_render.go: table rendering + presentation
  format helpers

18 funcs in, 18 funcs out (lossless). Verified by go build, go vet,
cmd/daemon tests, scoped golangci-lint, goimports.
